Remove pre-processor subpackage from anomalib
