+++ src/anomalib/data/utils/tiler.py
