--- src/anomalib/pre_processing/__init__.py
@@ -1,9 +1,0 @@
-"""Utilities for pre-processing the input before passing to the model."""
-
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .pre_process import PreProcessor
-from .tiler import ImageUpscaleMode, Tiler
-
-__all__ = ["PreProcessor", "Tiler", "ImageUpscaleMode"]
