--- src/anomalib/pre_processing/transforms/__init__.py
@@ -1,8 +1,0 @@
-"""Anomalib Data Transforms."""
-
-# Copyright (C) 2022 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .custom import Denormalize, ToNumpy
-
-__all__ = ["Denormalize", "ToNumpy"]
