+++ tests/unit/data/utils/__init__.py
@@ -1,4 +1,4 @@
-"""Pre-processing tests."""
+"""Unit tests - Data Utils."""
 
 # Copyright (C) 2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
