--- tests/unit/pre_processing/transforms/__init__.py
@@ -1,4 +1,0 @@
-"""Test transforms."""
-
-# Copyright (C) 2023 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
